source files push
Working on a synth for this

+++ ass_project/AES_Strip/aes_kernel.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
-void aes_main(uint8_t input[16], uint8_t output[16]);
+void aes_main(uint32_t input[16], uint32_t output[16]);
 void aes_decrypt(uint8_t input[16], uint8_t output[16]);
